Add values indexing while adding records and search by both indexed and unindexed fields.

+++ db-engine/src/database/structures.rs
@@ -4,6 +4,8 @@ pub mod filters;
 pub mod free_space;
 pub mod table_metadata;
 
+use bincode::Decode;
+use bincode::Encode;
 pub use dbtype::Type;
 pub use field::Field;
 pub use filters::Filter;
@@ -11,8 +13,8 @@ pub use filters::FilterOption;
 pub use free_space::FreeSpace;
 pub use table_metadata::TableMetadata;
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq, Encode, Decode, Hash)]
 pub struct DataPosition {
-    pub page: u16,
+    pub page: u64,
     pub cell: u16,
 }
